+++ CHANGELOG.markdown
@@ -6,6 +6,7 @@
 * Changed `_head`, `_tail`, `_init` and `_last` to traversals in `Data.Vector.Lens` and `Data.Vector.Generic.Lens`.
 * `Strict(strict)` is now a `Simple Iso`.
 * Added `withPrism` and `withIso`
+* Added `wrapped'` and `unwrapped'` for scenarios where you need the help with type inference.
 
 3.7.1.2 [maintenance release]
 -------